collision detection and optimization
collision and state now works on enemy. state not very important but eh. added typedefs for optimization in using smaller data types

+++ player.h
@@ -10,7 +10,7 @@
 class player
 {
 protected:
-	int player_speed;
+	int16 player_speed;
 public:
 	player();
 	~player();
